requirements.txt
Django>=3.1.1,<3.2.0
djangorestframework>=3.11.1,<3.12.0
psycopg2>=2.6
pytest-django>=4.1.0,<4.2.0
pytest-randomly>=3.4.1,<3.5.0
flake8>=3.8.3,<3.9.0
black>=20.8b1,<20.9
django-cors-headers>=3.6.0,<3.7.0
factory-boy>=2.12.0,<2.13.0